move sql db implementation into its own package
